+++ src/App.tsx
@@ -8,8 +8,10 @@ import Header from './components/Header/Header';
 import MainPage from './pages/MainPage/MainPage';
 import ProductPage from './pages/ProductPage/ProductPage';
 import CartPage from './pages/CartPage/CartPage';
+import SignupPage from './pages/SignupPage/SignupPage';
+import SignupStep from './pages/SignupPage/SignupStep';
 import NotFoundPage from './pages/NotFoundPage/NotFoundPage';
-import PATHS from './constants/paths';
+import { PATHS } from './constants/paths';
 
 function App() {
   return (
@@ -23,6 +25,9 @@ function App() {
             <Route path={PATHS.PRODUCT} element={<ProductPage />} />
             <Route path={PATHS.CART} element={<CartPage />} />
             <Route path={PATHS.DEFAULT} element={<NotFoundPage />} />
+            <Route path={PATHS.SIGNUP} element={<SignupPage />}>
+              <Route path={':stepId'} element={<SignupStep />} />
+            </Route>
           </Routes>
         </StyledContent>
       </ThemeProvider>
