+++ src/pages/SignupPage/Steps/FillInfoStep/FillInfoStep.types.ts
