+++ src/pages/SignupPage/Steps/ResultStep/ResultStep.stories.jsx
@@ -1,0 +1,12 @@
+import ResultStep from './ResultStep';
+
+export default {
+  title: 'Signup Step/ResultStep',
+  component: ResultStep,
+};
+
+function Template(args) {
+  return <ResultStep {...args} />;
+}
+
+export const Default = Template.bind({});
