+++ src/pages/SignupPage/Steps/ResultStep/ResultStep.styled.ts
@@ -1,14 +1,5 @@
 import styled from 'styled-components';
 
-export const PageBox = styled.div`
-  width: 600px;
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-  justify-content: center;
-  align-items: center;
-`;
-
 export const ResultBox = styled.div`
   display: flex;
   flex-direction: column;
