+++ src/pages/SignupPage/Steps/TermsStep/TermsStep.stories.jsx
@@ -1,0 +1,12 @@
+import TermsStep from './TermsStep';
+
+export default {
+  title: 'Signup Step/TermsStep',
+  component: TermsStep,
+};
+
+function Template(args) {
+  return <TermsStep {...args} />;
+}
+
+export const Default = Template.bind({});
