+++ src/pages/SignupPage/Steps/TermsStep/TermsStep.styled.ts
@@ -1,14 +1,5 @@
 import styled from 'styled-components';
 
-export const PageBox = styled.div`
-  width: 600px;
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-  justify-content: center;
-  align-items: center;
-`;
-
 export const TermsBox = styled.div`
   width: 100%;
   display: flex;
