--- src/pages/SignupResultPage/SignupResultPage.stories.jsx
@@ -1,12 +1,0 @@
-import SignupResultPage from './SignupResultPage';
-
-export default {
-  title: 'Page/SignupResultPage',
-  component: SignupResultPage,
-};
-
-function Template(args) {
-  return <SignupResultPage {...args} />;
-}
-
-export const Default = Template.bind({});
