--- src/pages/TermsPage/TermsPage.stories.jsx
@@ -1,12 +1,0 @@
-import TermsPage from './TermsPage';
-
-export default {
-  title: 'Page/TermsPage',
-  component: TermsPage,
-};
-
-function Template(args) {
-  return <TermsPage {...args} />;
-}
-
-export const Default = Template.bind({});
